+++ src/components/Router/StepComponentMap.js
@@ -2,14 +2,14 @@ import { h } from 'preact'
 
 import Welcome from '../Welcome'
 import {SelectPoADocument, SelectIdentityDocument} from '../Select'
-import {FrontDocument, BackDocument, Selfie, Liveness, PoADocument } from '../Capture'
-import {DocumentFrontConfirm, DocumentBackConfirm, FaceConfirm, LivenessConfirm} from '../Confirm'
+import {FrontDocument, BackDocument, Selfie, Video, PoADocument } from '../Capture'
+import {DocumentFrontConfirm, DocumentBackConfirm, SelfieConfirm, VideoConfirm} from '../Confirm'
 import Complete from '../Complete'
 import MobileFlow from '../crossDevice/MobileFlow'
 import CrossDeviceLink from '../crossDevice/CrossDeviceLink'
 import ClientSuccess from '../crossDevice/ClientSuccess'
 import CrossDeviceIntro from '../crossDevice/Intro'
-import LivenessIntro from '../Liveness/Intro'
+import VideoIntro from '../Video/Intro'
 import PoAIntro from '../ProofOfAddress/PoAIntro'
 
 export const componentsList = ({flow, documentType, steps, mobileFlow}) => {
@@ -26,7 +26,7 @@ const hasCompleteStep = (steps) => steps.some(isComplete)
 const clientCaptureSteps = (steps) =>
   hasCompleteStep(steps) ? steps : [...steps, {type: 'complete'}]
 
-const shouldUseLiveness = steps => {
+const shouldUseVideo = steps => {
   const { options: faceOptions } = Array.find(steps, ({ type }) => type === 'face') || {}
   return (faceOptions || {}).requestedVariant === 'video' && window.MediaRecorder
 }
@@ -36,9 +36,9 @@ const captureStepsComponents = (documentType, mobileFlow, steps) => {
 
   return {
     welcome: () => [Welcome],
-    face: () => shouldUseLiveness(steps) ?
-        [LivenessIntro, Liveness, LivenessConfirm] :
-        [Selfie, FaceConfirm],
+    face: () => shouldUseVideo(steps) ?
+        [VideoIntro, Video, VideoConfirm] :
+        [Selfie, SelfieConfirm],
     document: () => createIdentityDocumentComponents(documentType),
     poa: () => [PoAIntro, SelectPoADocument, PoADocument, DocumentFrontConfirm],
     complete: () => complete
